Implemented Music Player Screen Logic

+++ app/src/main/java/com/xeniac/chillclub/core/data/utils/MathHelper.kt
@@ -3,4 +3,4 @@ package com.xeniac.chillclub.core.data.utils
 fun Int.convertToPercentage(
     minValue: Int,
     maxValue: Int
-): Float = (this - minValue) / (maxValue - minValue) * 100f
+): Float = (this - minValue).toFloat() / (maxValue - minValue).toFloat() * 100f
